backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
python-dotenv==1.0.0
pytest==7.4.2
pytest-flask==1.2.0
email-validator==2.0.0